ml-engine/requirements.txt
psycopg2-binary==2.9.10
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
